+++ src/ModuarkServiceProvider.php
@@ -8,6 +8,7 @@ use Cluion\Moduark\Analysis\ArchitectureCheck;
 use Cluion\Moduark\Analysis\ArchitectureChecker;
 use Cluion\Moduark\Analysis\Boundary\ConventionPublicApi;
 use Cluion\Moduark\Analysis\Boundary\PublicApi;
+use Cluion\Moduark\Analysis\Export\GithubCheckReportExporter;
 use Cluion\Moduark\Analysis\Export\JsonCheckReportExporter;
 use Cluion\Moduark\Analysis\RuleRunner;
 use Cluion\Moduark\Analysis\Rules\AdapterBoundariesRule;
@@ -92,6 +93,7 @@ final class ModuarkServiceProvider extends ServiceProvider
         );
         $this->app->singleton(ModuleMetadataCompiler::class);
         $this->app->singleton(SourceIndexBuilder::class);
+        $this->app->singleton(GithubCheckReportExporter::class);
         $this->app->singleton(JsonCheckReportExporter::class);
         $this->app->singleton(CapabilityGraphBuilder::class);
         $this->app->singleton(CombinedGraphBuilder::class);